Godot4 stuff1 (#57)
From Godot 4.0
* Introduce Callable and convert ObjectID to a proper class.
* Callable/Signal/StringName are now basic Variant types

Removed/Changed
- visual scripting 
- complex expression evaluator from editor number fields
* big Variant class cleanup
+ Update binding generator to correctly handle generating a signal (Event+Delegate)
* Windows build fixes.
* Most Variant conversion operators around the codebase were converted to templated .as<T> functions <- this was needed to find a gnarly bug and could be reverted to explicit conversions.
* some generic clang-tidy fixups.

+++ modules/mono/editor/GodotTools/GodotTools/Internals/ScriptClassParser.cs
@@ -13,12 +13,12 @@ namespace GodotTools.Internals
             public string Name { get; }
             public string Namespace { get; }
             public bool Nested { get; }
-            public int BaseCount { get; }
+            public long BaseCount { get; }
 
             public string SearchName => Nested ?
                 Name.Substring(Name.LastIndexOf(".", StringComparison.Ordinal) + 1) :
                 Name;
-            public ClassDecl(string name, string @namespace, bool nested, int baseCount)
+            public ClassDecl(string name, string @namespace, bool nested, long baseCount)
             {
                 Name = name;
                 Namespace = @namespace;
@@ -48,7 +48,7 @@ namespace GodotTools.Internals
                     (string)classDeclDict["name"],
                     (string)classDeclDict["namespace"],
                     (bool)classDeclDict["nested"],
-                    (int)classDeclDict["base_count"]
+                    (long)classDeclDict["base_count"]
                 ));
             }
